refactor(schematics): move helper schematics to subdirectory
