+++ schematics/src/helpers/customized-copy/schema.json
