+++ schematics/src/helpers/lazy-components/factory.ts
