+++ schematics/src/helpers/lazy-components/schema.json
