+++ schematics/src/helpers/move-component/schema.json
